+++ src/i18n/en.ts
@@ -355,6 +355,8 @@ const en = {
       labelInactive: 'Detected / Closed',
     },
 
+    doa: '%s will never be triggered as currently configured. Did you forget to enable webhooks or add a trigger schedule?',
+
     humidity: '%s humidity is %d%', // accessory name, number
 
     leak: {