+++ src/i18n/missing/de.ts
@@ -229,6 +229,8 @@ const de = {
       labelInactive: 'Detected / Closed',
     },
 
+    doa: '%s will never be triggered as currently configured. Did you mean to enable webhooks or add a schedule?',
+
     humidity: '%s humidity is %d%', // accessory name, number
 
     leak: {