+++ src/i18n/missing/vi.ts
@@ -36,6 +36,8 @@ const vi = {
       labelInactive: 'Detected / Closed',
     },
 
+    doa: '%s will never be triggered as currently configured. Did you mean to enable webhooks or add a schedule?',
+
     leak: {
       labelActive: 'Detected',
       labelInactive: 'Not Detected',